+++ app/src/main/java/fr/iut/cascade/GameActivity.java
@@ -1,10 +1,19 @@
 package fr.iut.cascade;
 
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.MotionEvent;
+import android.view.View;
 import android.view.WindowManager;
+import android.widget.ImageView;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import org.w3c.dom.Text;
+
+import java.util.ArrayList;
 
 import fr.iut.cascade.game.Grid;
 import fr.iut.cascade.game.listeners.GridEventListener;
@@ -37,6 +46,7 @@ public class GameActivity extends AppCompatActivity {
     private int difficulty;
 
     public final static String SHARED_PREFERENCES_SCOREBOARD_NAME = "scoreboard";
+    public final static String LAST_SCORE = "last_score";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -45,6 +55,10 @@ public class GameActivity extends AppCompatActivity {
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_game);
+
+        initButtons();
+        findViewById(R.id.end_layout).setVisibility(View.INVISIBLE
+        );
 
         difficulty = 1;
         // Recovery of the difficulty chosen
@@ -74,11 +88,113 @@ public class GameActivity extends AppCompatActivity {
             public void onGameFinished(Grid grid) {
                 int score = grid.getScore();
                 String score_str = Integer.toString(grid.getScore());
-                SettingsUtil.saveScore(score, grid.getDifficulty(), SHARED_PREFERENCES_SCOREBOARD_NAME, getApplicationContext());
-                Toast.makeText(getBaseContext(), getString(R.string.score_label) + " " + score_str, Toast.LENGTH_LONG).show();
-                grid.reset();
-                ((TextView) findViewById(R.id.scoreValue)).setText(score_str);
+
+                // Save the score and get the place
+                int place = SettingsUtil.saveScore(score, grid.getDifficulty(), SHARED_PREFERENCES_SCOREBOARD_NAME, getApplicationContext());
+                String place_str = Integer.toString(place);
+
+                // Set up the end screen
+                findViewById(R.id.end_layout).setVisibility(View.VISIBLE);
+                TextView end_score = findViewById(R.id.end_screen_score);
+                TextView end_place = findViewById(R.id.end_screen_place);
+                TextView end_combo = findViewById(R.id.end_screen_combo);
+                TextView end_clicks = findViewById(R.id.end_screen_clicks);
+                ImageView end_difficulty = findViewById(R.id.end_screen_difficulty);
+
+                if(place == -1){
+                    end_place.setText(getString(R.string.not_record));
+                } else if(place == 1){
+                    end_place.setText(getString(R.string.best_record));
+                } else{
+                    String message = getString(R.string.record) + place_str;
+                    end_place.setText(message);
+                }
+
+                end_score.setText(score_str);
+
+                String max_combo = getString(R.string.max_combo) + " " + Integer.toString(grid.getBestCombo()) + " (" + Integer.toString(grid.getBestComboScore()) + ")";
+                end_combo.setText(max_combo);
+
+                String total_clicks = getString(R.string.total_clicks) + " " +Integer.toString(grid.getTotalClicks());
+                end_clicks.setText(total_clicks);
+
+                switch (difficulty){
+                    case 1 :
+                        end_difficulty.setImageResource(R.mipmap.dif_1);
+                        break;
+                    case 2:
+                        end_difficulty.setImageResource(R.mipmap.dif_2);
+                        break;
+                    case 3:
+                        end_difficulty.setImageResource(R.mipmap.dif_3);
+                        break;
+                    case 4:
+                        end_difficulty.setImageResource(R.mipmap.dif_4);
+                        break;
+                    default:
+                }
             }
         });
+    }
+
+    /**
+     * Called whenever a button is pressed
+     * @param view view that sent the method
+     */
+    public void onButtonClick(View view) {
+        switch (view.getId()){
+            case R.id.scoreboard_button:
+                Intent scoreboardIntent = new Intent(this, ScoreboardActivity.class);
+                scoreboardIntent.putExtra(MainActivity.DIFFICULTY, this.difficulty);
+                scoreboardIntent.putExtra(LAST_SCORE, ((Grid) findViewById(R.id.grid)).getScore());
+                startActivity(scoreboardIntent);
+                break;
+            case R.id.restart_button:
+                Grid grid = findViewById(R.id.grid);
+
+                // Restart the game
+                grid.reset();
+                String score_str = Integer.toString(grid.getScore());
+                ((TextView) findViewById(R.id.scoreValue)).setText(score_str);
+
+                // Remove the end screen
+                findViewById(R.id.end_layout).setVisibility(View.INVISIBLE);
+
+                break;
+            default:
+                break;
+        }
+    }
+
+    /**
+     * Init the buttons dynamic (pushed / released)
+     */
+    private void initButtons(){
+        ArrayList<int[]> buttons = new ArrayList<>();
+        buttons.add(new int[]{R.id.restart_button, R.mipmap.restart_default, R.mipmap.restart_pushed});
+        buttons.add(new int[]{R.id.scoreboard_button, R.mipmap.scoreboard_default, R.mipmap.scoreboard_pushed});
+
+        for (int[] values : buttons) {
+            int button_id = values[0];
+            final int button_res_default = values[1];
+            final int button_res_pushed = values[2];
+
+            // Animate the buttons
+            findViewById(button_id).setOnTouchListener(new View.OnTouchListener() {
+                @Override
+                public boolean onTouch(View view, MotionEvent motionEvent) {
+                    // Pressed
+                    if(motionEvent.getAction() == MotionEvent.ACTION_DOWN){
+                        ((ImageView) view).setImageResource(button_res_pushed);
+                    }
+                    // Released
+                    if(motionEvent.getAction() == MotionEvent.ACTION_UP){
+                        ((ImageView) view).setImageResource(button_res_default);
+                        view.performClick();
+                    }
+                    return true;
+                }
+            });
+        }
     }
 }