+++ CHANGELOG.md
@@ -19,8 +19,8 @@ This changelog includes only user-sided updates. It indicates the advancement fo
 
 ### Version 2 "User friendly interface" - Release date : ??/??/??
 - ✔ Clean overall theme
-- 🏳 End screen with useful information (score rank, number of clicks, ...)
+- ✔ End screen with useful information (score place, score total, restart option, best combo, number of clicks)
 - 🏳️ Animations
 - 🏳️ Visual effects on touch
 - 🏳 Sound
-- 🏳 Statistics
+- 🏳 Statistics on the scoreboard